% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_onrg}
\alias{predict_onrg}
\title{Read and analyze univariate time series data for EU datathon}
\usage{
predict_onrg(json_data, time="years", amount="values", order=NULL,
 prediction_steps= 5, tojson= TRUE)
}
\arguments{
\item{json_data}{The json text or link with the EU Energy data frame or matrix}

\item{time}{Time label of the json time series data}

\item{amount}{Amount label of the json time series data}

\item{order}{An integer vector of length 3 specifying the order of the Arima model}

\item{prediction_steps}{The number of prediction steps}

\item{tojson}{If FALSE returns a list object}
}
\value{
A json string with the resulted parameters of the ts.analysis function.
}
\description{
Extract and analyze univariate time series data for EU datathon
}
\details{
This function is used to predict eu energy statistics/market data in OpEnergy application.
}
\author{
Kleanthis Koupidis
}
