% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_target.R
\name{predict_res_target}
\alias{predict_res_target}
\title{Predictions}
\usage{
predict_res_target(json, targets2020="target", forward_steps= 5, toJSON= TRUE)
}
\arguments{
\item{json}{The json text or link with the EU Energy data frame or matrix}

\item{targets2020}{The 2020 targets.}

\item{forward_steps}{The prediction steps}

\item{toJSON}{If FALSE returns a list object}
}
\value{
A json string with the necessary parameters to visualize the results in OpEnergy application.
}
\description{
Extract and analyze univariate time series data in OpEnergy application http://openergy.okfn.gr
}
\details{
This function is used to predict eu energy market data used in OpEnergy application
}
\seealso{
\code{\link[TimeSeries.OBeu]{ts.analysis}}
}
\author{
Kleanthis Koupidis
}
